switch deps to sdk-utils
