+++ packages/core/package.json
@@ -30,7 +30,6 @@
     "dependencies": {
         "@flexn/plugins": "1.0.9",
         "ajv": "7.0.0-beta.0",
-        "axios": "0.21.2",
         "color-string": "1.9.0",
         "deepmerge": "3.2.0",
         "detect-port": "1.3.0",
