--- packages/engine-rn-electron/types.d.ts
@@ -1,1 +1,0 @@
-declare module 'better-opn'
