--- packages/engine-rn-next/types.d.ts
@@ -1,1 +1,0 @@
-declare module 'better-opn'
