+++ packages/engine-rn-web/types.d.ts
@@ -1,2 +1,1 @@
-declare module 'better-opn'
 declare module 'ip'
