+++ packages/sdk-utils/package.json
@@ -28,7 +28,9 @@
         "lint": "eslint . --ext .ts",
         "watch": "tsc --watch --preserveWatchOutput"
     },
-    "dependencies": {},
+    "dependencies": {
+        "better-opn": "1.0.0"
+    },
     "peerDependencies": {
         "@rnv/core": "^1.0.0-canary.7"
     },
