+++ packages/sdk-utils/src/index.ts
@@ -1,1 +1,66 @@
-export default {};
+import { DEFAULTS, RnvContext, getConfigProp, logTask } from '@rnv/core';
+import axios from 'axios';
+import open from 'better-opn';
+
+export const getValidLocalhost = (value: string, localhost: string) => {
+    if (!value) return localhost;
+    switch (value) {
+        case 'localhost':
+        case '0.0.0.0':
+        case '127.0.0.1':
+            return localhost;
+        default:
+            return value;
+    }
+};
+
+export const openBrowser = open;
+
+export const getDevServerHost = (c: RnvContext) => {
+    const devServerHostOrig = getConfigProp(c, c.platform, 'devServerHost');
+
+    const devServerHostFixed = devServerHostOrig
+        ? getValidLocalhost(devServerHostOrig, c.runtime.localhost || DEFAULTS.devServerHost)
+        : DEFAULTS.devServerHost;
+
+    return devServerHostFixed;
+};
+
+export const waitForHost = async (c: RnvContext, suffix = 'assets/bundle.js') => {
+    logTask('waitForHost', `port:${c.runtime.port}`);
+    let attempts = 0;
+    const maxAttempts = 10;
+    const CHECK_INTEVAL = 2000;
+    // const spinner = ora('Waiting for webpack to finish...').start();
+
+    const devServerHost = getDevServerHost(c);
+    const url = `http://${devServerHost}:${c.runtime.port}/${suffix}`;
+
+    return new Promise((resolve, reject) => {
+        const interval = setInterval(() => {
+            axios
+                .get(url)
+                .then((res) => {
+                    if (res.status === 200) {
+                        clearInterval(interval);
+                        // spinner.succeed();
+                        return resolve(true);
+                    }
+                    attempts++;
+                    if (attempts === maxAttempts) {
+                        clearInterval(interval);
+                        // spinner.fail('Can\'t connect to webpack. Try restarting it.');
+                        return reject(`Can't connect to host ${url}. Try restarting it.`);
+                    }
+                })
+                .catch(() => {
+                    attempts++;
+                    if (attempts > maxAttempts) {
+                        clearInterval(interval);
+                        // spinner.fail('Can\'t connect to webpack. Try restarting it.');
+                        return reject(`Can't connect to host ${url}. Try restarting it.`);
+                    }
+                });
+        }, CHECK_INTEVAL);
+    });
+};
