+++ packages/sdk-utils/types.d.ts
@@ -1,2 +1,1 @@
-declare module 'command-exists'
 declare module 'better-opn'
